requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
jinja2==3.1.4
python-multipart==0.0.12
requests==2.32.3
openpyxl==3.1.5
python-dotenv==1.0.1
pillow==10.4.0
pytesseract==0.3.13
python-dateutil==2.9.0
apscheduler==3.10.4
pydantic==2.9.2
pypdf==5.0.1
alembic==1.13.3
psycopg2-binary==2.9.10
bcrypt==4.2.0
pyjwt==2.9.0
